@@ -5,25 +5,32 @@ import napari
 import unittest
 from unittest.mock import MagicMock, patch, call
 from pathlib import Path
+import pytest
 from napari.layers import Shapes
-from livecell_tracker import sample_data
-from livecell_tracker.core.napari_visualizer import NapariVisualizer
-from livecell_tracker.segment.utils import prep_scs_from_mask_dataset
-from livecell_tracker.core import (
+from livecellx import sample_data
+from livecellx.core.napari_visualizer import NapariVisualizer
+from livecellx.core.io_sc import prep_scs_from_mask_dataset
+from livecellx.core import (
     SingleCellTrajectory,
     SingleCellStatic,
     SingleCellTrajectoryCollection,
 )
-from livecell_tracker.track.sort_tracker_utils import (
+from livecellx.track.sort_tracker_utils import (
     gen_SORT_detections_input_from_contours,
     update_traj_collection_by_SORT_tracker_detection,
     track_SORT_bbox_from_contours,
     track_SORT_bbox_from_scs,
 )
-from livecell_tracker.core.sct_operator import SctOperator
+from livecellx.core.sct_operator import SctOperator
 
 
 class SctOperatorTest(unittest.TestCase):
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+        # Skip the entire test class
+        pytest.skip("Skipping SctOperatorTest")
+
     @classmethod
     def setUpClass(cls):
         dic_dataset, mask_dataset = sample_data.tutorial_three_image_sys()
@@ -39,9 +46,9 @@ class SctOperatorTest(unittest.TestCase):
         )
         self.sct_operator = SctOperator(self.traj_collection, self.shape_layer, self.viewer)
 
-        self.sample_dir = ".\\test_sample_dir"
-        if not os.path.exists(self.sample_dir):
-            os.makedirs(self.sample_dir)
+        self.sample_dir = Path("./test_sample_dir")
+        if not self.sample_dir.exists():
+            self.sample_dir.mkdir(parents=True)
 
     def test_delete_selected_sct(self):
         # Given: Initial state
@@ -59,6 +66,9 @@ class SctOperatorTest(unittest.TestCase):
 
         self.sct_operator.select_info = select_info_list
 
+        all_track_ids = self.traj_collection.get_all_tids()
+        all_track_ids.remove(track_id_to_delete)
+
         # When: Deleting selected trajectory
         self.sct_operator.delete_selected_sct()
 
@@ -69,9 +79,11 @@ class SctOperatorTest(unittest.TestCase):
         # 2. Check if the selection info is cleared
         self.assertEqual(self.sct_operator.select_info, [])  # Assuming select_info is cleared after deletion
 
-        # 3. Check if the shape related to the deleted trajectory has been removed from the shape_layer
+        # 3. Check if the only shape related to the deleted trajectory has been removed from the shape_layer
         shape_track_ids = self.sct_operator.shape_layer.properties["track_id"]
         self.assertNotIn(track_id_to_delete, shape_track_ids)
+        for tid in all_track_ids:
+            self.assertIn(tid, shape_track_ids)
 
     def test_save_annotations(self):
         # Given: Setup data
@@ -104,11 +116,11 @@ class SctOperatorTest(unittest.TestCase):
                 self.assertTrue(isinstance(data, list))
 
                 # 2. Verify each item in the list is a dictionary representing a SingleCellStatic instance
-                for item in data:
-                    self.assertTrue(isinstance(item, dict))
+                for sc_dict in data:
+                    self.assertTrue(isinstance(sc_dict, dict))
                     # Example checks for the expected keys in each dictionary
                     for key in ["id", "timeframe", "bbox", "contour", "meta", "dataset_json_dir"]:
-                        self.assertIn(key, item)
+                        self.assertIn(key, sc_dict)
 
     def tearDown(self):
         # Cleanup: Delete all files and directories recursively
