@@ -2,11 +2,11 @@ import json
 import unittest
 from pathlib import Path
 import numpy as np
-from livecell_tracker import sample_data
-from livecell_tracker.core.sc_key_manager import SingleCellMetaKeyManager as SCKM
-from livecell_tracker.segment.utils import prep_scs_from_mask_dataset
-from livecell_tracker.core.datasets import LiveCellImageDataset
-from livecell_tracker.core import (
+from livecellx import sample_data
+from livecellx.core.sc_key_manager import SingleCellMetaKeyManager as SCKM
+from livecellx.core.io_sc import prep_scs_from_mask_dataset
+from livecellx.core.datasets import LiveCellImageDataset
+from livecellx.core import (
     SingleCellTrajectory,
     SingleCellStatic,
     SingleCellTrajectoryCollection,
