@@ -1,3 +1,3 @@
-# from livecell_tracker.model_zoo.segmentation.eval_csn import evaluate_sample_v3_underseg, match_label_mask_by_iou
+# from livecellx.model_zoo.segmentation.eval_csn import evaluate_sample_v3_underseg, match_label_mask_by_iou
 
 # TODO: test functions used in evaluation
